requirements.txt
allure-combine==1.0.11
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deepdiff==7.0.1
et-xmlfile==1.1.0
exceptiongroup==1.2.2
Flask==3.0.3
idna==3.7
importlib_metadata==8.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpath==0.82.2
MarkupSafe==2.1.5
numpy==2.0.0
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
pluggy==1.5.0
cryptography
pycryptodome==3.20.0
pycryptodomex
PyMySQL==1.1.1
pytest==8.0.2
pytest-rerunfailures==14.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
tomli==2.0.1
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
